+++ Modules/Summary.lua
@@ -14,8 +14,18 @@ function Module:OnInitialize()
 end
 
 function Module:OnEnable()
-    self:SecureHookScript(CharacterStatsPane.ItemLevelFrame.Corruption, 'OnEnter', 'SummaryEnter')
-    self:SecureHookScript(CharacterStatsPane.ItemLevelFrame.Corruption, 'OnLeave', 'SummaryLeave')
+    if IsAddOnLoaded("Narcissus") then
+        self:SecureHook('PaperDollFrame_UpdateCorruptedItemGlows', function(enter)
+            if enter then
+                Module:SummaryEnter()
+            else
+                Module:SummaryLeave()
+            end
+        end)
+    else
+        self:SecureHookScript(CharacterStatsPane.ItemLevelFrame.Corruption, 'OnEnter', 'SummaryEnter')
+        self:SecureHookScript(CharacterStatsPane.ItemLevelFrame.Corruption, 'OnLeave', 'SummaryLeave')
+    end
 end
 
 function Module:OnLoad(event, ...)